[0003] Backup Commit
- backup before starting with following idea:
Source-filter connection tree is modular. Each user-unique source has a
list of filter(s) associated. The list of sources and targets build the
set. Associated filters are derived from the sources. The list of
targets is static and can't be changed. Target-cateogization would be
achieved through different sets.
The UI groups sets that all together point to a single filter/set of
filters or no filter at all (directly target set).
A filter consists of a list of rules that are applied in sequence as
such:
- [object (file, folder)] [matches/starts with/ends with/etc] [pattern]
- [object (file, folder)] [matches/mismatches] [regex]

Filters can be chained up (as nodes) in open trees so different sources
can traverse and enter a tree at different points. A tree always ends in
one set of targets.

Iterative thoughts:
- Sources might not be grouped but stand alone...
- sources and all filters only save *outgoing* connections. Incoming
ones are derived on their inbound objects (as outgoing connections
there).