[0003] End of Day Commit
2013-05-25 - 27
- implemented save-methodology for nodes
	- set-table on db has a column "gui_data" that saves a dictionary with
a namespace for the nodes, ids, etc. and their x/y (center) coordinates.
That's all for now, will be able to add data later on easily.
	- added save button
	- added timeout signals that exectue the save_to_db after a 1s timeout
if the modified_signal has been triggered. During that time the save
button gets enabled making it possible to click it manually (If you're
really in a hurry)